fetch data from api

+++ lib/home/data/models/home_model/datum.dart
@@ -7,7 +7,7 @@ final  String description;
 final  String location;
 final  DateTime postedOn;
 final  int salaryTypeId;
-final  int salary;
+final  double salary;
 final  int jobTypeTd;
